Redo the new keepalive server-side tests to use nginx as the test server
This allows us to get rid of the custom EventMachine server we were
using in this case, and just use the same OpenResty test server as we do
for all the other tests.

In order to do so, we need to move the test server to a separate nginx
master process, rather than embedding it inside our existing API
Umbrella nginx server when run in the test environment. But this is
probably a better approach anyway, so the test server acts more like an
independent API.

By splitting the test server to it's own nginx process, now we can
reliably, and more easily use nginx's various status variables to track
how many open and idle connections it has against it for checking the
keep-alive behavior. We couldn't really use these variables when the
test server was inside the same nginx process as API Umbrella, since the
counts were skewed by all the other connections API Umbrella may be
holding open internally.