Improve testing resiliency in case certain tests take longer.
Our previous approach to checking cache timing was dependent on the
requests all happening within 1 second of each other. That could fail if
we were running parallel tests and other processes were contending for
access. Since all we really care about is whether the Age header
increases from the previous amount, change the test so they're not as
sensitive to exact timing.