Fix possible errors in capybara tests caused by fixed nav header.
Depending on the order of tests,
Test::AdminUi::TestDatatables#test_spinner_on_server_side_loads could
reliably start failing, if there was ApiUser data in the system before
it was called. The issued stemmed from the test trying to click on the
table header, but then it couldn't click on it, since it reported that
the navbar was on top of it. I think this has to do with Poltergeist's
scrolling logic when it needs to scroll an element into view. With a
fixed, floating header, then it's logic wasn't quite right, since it
scrolled it technically into view, but underneath the fixed header.

Since we don't really care about the fixed nav header for Capybara
tests, disable it's fixed positioning so it just flow on the page
normally.