Change how we fetch the requests logs from Elasticsearch in tests.
Instead of querying the database based on the unique query parameter,
we'll instead fetch records directly based on the request ID.

This begins to reduce the dependency on logging the individual request
query params for the test suite. We've disabled logging this data by
default (see a432b220), but we had kept it enabled for the test suite.
This removes one of the bigger barriers to disabling it in the test
environment too (although, there's still some other parts of the test
suite that depend on it, so those still need to be revisited).

This also partially stemmed from some odd test failures and debugging.
On very rare occasions, the data was getting indexed into elasticsearch,
but wait_for_log was failing to find them. The culprit appeared like the
analyzers weren't working properly, since the queries for the
unique_query_id were only working if the query was downcased. But if it
got in this funky state, restarting elasticsearch would return things to
normal where queries worked in the expected case-insensitive fashion. I
think this behavior might be fixed as part of the various mapping and
analyzer cleanup in Elasticsearch 2. It was quite difficult to reproduce
this issue, so shifting to this ID-based query approach should also
hopefully eliminate these rare edge case issues in the test suite too.