Change TrafficServer log format to include more details.
This is based on the Netscape log format, so it's more similar to our
nginx logs. But this also logs further details to help in debugging
(right now to debug intermittent test failures, but since this logs is
mainly for debugging purposes on production too, the extra details seem
relevant there too).