Change IP used in test due to changes in IP database.
For our test of IPs with only countries, but no city and region, our
Singapore test IP now returns a city (I think this just changed in
GeoIP's database released today). So switch to another IP that only
returns a country code, but no city or region to test something similar
(the new IP isn't actually a valid country code, it represent anonymous
proxies, but it's the only IP I could find that returns data with only a
country code, which is what we really want to test here).