Wait for all remote data fetches to complete before loading forms.
In some of our forms, we need to also fetch other data from the server
to generate option lists (for example, the list of admin groups
available). However, previously we were fetching that other data
asynchronously after the form had loaded. This changes things so we
fetch the record and any associated data at the same time, and wait to
render the form until both are complete.

This has the benefit of showing the "loading" spinner until all data is
fetched for displaying the form (rather than being partially loaded).

This also seems to finally help fix some random race conditions that
could crop up in our Capybara tests, since the asynchronous ajax
could lead to a partially loaded form, or the ajax requests could be
left hanging in the background if we navigated away without waiting for
everything to complete (which could throw some javascript errors).