Improve asset precompiling when switching between branches in dev.
Skip reading the runtime_config.yml file when running asset
precompiling, and just fallback to the default.yml settings for asset
precompile. This helps prevent issues when switching between branches in
a development environment, since the last runtime_config.yml file might
not be associated with the current branch, and therefore might not have
all the required data in it.

Also skip devise routes in precompile mode to speed things up a tad.