Ensure elasticsearch plugins directory exists.
Fixes #318.

This directory needs to exist for elasticsearch to startup, but it's not
part of the installation. Elasticsearch will normally create this on
startup if it has permissions to, but after our installation, the parent
directory is owned by root, so that's why we need to pre-create this for
our package installations.