Switch to using LD_LIBRARY_PATH over RPATHs.
This should fix https://github.com/NREL/api-umbrella/issues/319

By ensuring consistent LD_LIBRARY_PATH usage across all libraries and
binaries, we should be able to sidestep the development staging vs
install location issue in hopefully a cleaner, more consistent way. This
also makes the installed files more relocatable, since there's aren't
static paths compiled in.

This does require LD_LIBRARY_PATH to be setup properly before executing
scripts. The existing api-umbrella binary was already responsible for
setting other parts of the environment, so this is in line with that.
But to help make the environment setup more reusable for one-off scripts
and debugging purposes, we've abstracted the environment setup into a
couple of new scripts (api-umbrella-env, which can be sourced, and
api-umbrella-exec, which can be used to wrap other commands).