Add matched API backend ID to analytics logs.
See https://github.com/NREL/api-umbrella/issues/252

This logs the API backend ID that was matched while serving the API
request, along with the ID of the more specific "url_matches" record
within that API.

These details might serve as a more efficient way to lookup logs that a
user has permissions to view based on IDs (versus the host and path
prefix based current approach). Although, switching to use this might be
tricky given our current approach to admin permissions (since IDs are a
little less flexible and we have to deal with historical IDs). But in
any case, let's start logging this so we can explore this and allow for
other use cases that have requested this functionality.