Fix authentication request loop in admin when multiple windows open.
If multiple windows or tabs were open in the admin, there would be a
constant stream of requests to the "/admin/auth" endpoint. This is
because each new request contained a new CSRF token value (since it's
different on every server-side request). This in turn led to the local
storage being updated with new content, which then triggered an infinite
loop between the various tabs/windows fighting over what was the correct
data.

To sidestep this issue altogether, we'll no longer return the CSRF token
in the auth endpoint, and shift the admin authentication to use the
admin auth token value, which was already supported on all the API
endpoints. The only endpoint where we needed to actually handle CSRF was
the logout endpoint, so that's been updated to work with the admin token
value.