Increase default memory for "active_config" shared dict.
This is related to https://github.com/18F/api.data.gov/issues/385

This should allow the default configuration to publish somewhere in the
neighborhood of 750-1500 API backends (up from 150-300) without running
into potential issues.

We still need to address the more underlying issue so the shard dict
operates in a safer manner (eg, by using set_safe as described in the
github issue).