Fix capistrano deployments from master.
Capistrano-based deploys (to overlay changes from master on top of an
existing package installation) had become broken somewhere along the way
to the v0.14 release. This fixes the deployments by refactoring the
deployments to leverage more of the normal build process, so our
existing build process for gems/lua/admin-ui can be reused.

https://github.com/NREL/api-umbrella/issues/351