Add more tests for all the API "index" methods. Small search change.
This adds a basic test suite across all the API "index" methods that
are used in DataTables to ensure they behave as DataTables expects
(pagination, searching, etc).

To make testing easier, also add searching for the "_id" field to all
the searching.

Also adjust "_id" searching to only match the full value, rather than
wildcards (since I'm not sure we would ever really need wildcard
searches for IDs, and this will make searches a bit easier under the
Postgres stuff we're exploring with the explicit UUID type).

To make testing easier, also adjust how the datatables "order" param
gets parsed to allow for missing indexes (even though this shouldn't
really be happening in the real world). But this makes it easier to
test to ensure indexes are looped over in numeric order (so we can send
in "2" and "12" to test this).