Add pgcrypto and generate default UUID IDs in postgres.
This is mainly to keep our test suite more consistent with the previous
behavior which didn't have IDs in the factories (this was leading to
some test issues in cases where there were now duplicate IDs when
factory data was re-used). But this might be a better approach anyway
(as long as we're okay with the pgcrypto dependency, but I think it's
pretty common).