Ensure api backends have gaps in sort_orders when relying on default.
When relying on the database default, as the test factories do, ensure
there are gaps in the sort_order. This prevents our after-commit logic
in the api to ensure unique orders from getting stuck in a tight loop
(none of this is particularly ideal, though--ideally we'd have a real
uniqueness constraint on the table, but we might be able to revisit this
completely and get rid of sort orders at some point).