Handle 404s in Lapis app better. Add tests around Lapis error handling.
Return routing errors as 404s instead of 500 errors. Add testing around
both the 404 and 500 error handling behavior in Lapis. Remove the
error backtrace from the HTTP Headers for 500 errors (it turns out this
was only happening in the test environment, so this should not have been
a production issue, but remove it in the test environment too so that we
can better test this behavior).