Fix admin session cookie being stripped from certain admin requests.
The problem with the previous logic was that the "_api_umbrella_session"
cookie was being appended to the global config list of cookies to strip
anytime a non-web backend was seen. This config list was then persisted
across requests, so this was then affecting web backend requests on
subsequent requests.

This was specifically breaking the CSV download for the admin drilldown
reports (since this was one of the few requests routed through the API
backend that needed the admin session cookie, instead of relying on the
admin token HTTP header).