More testing and API fixes from postgres/lapis migration.
Notably, this changes how we return the default null value in the lapis
APIs. The previous "or null" approach would fail to return "false"
values, so use a helper function to properly deal with this (technically
this is only needed for boolean types, but it seems to make sense to
standardize on using it everywhere).