Fix failing COUNT queries when strings contained question marks.
The COUNTY query in the model_ext validate_uniqueness could fail if the
strings involved happened to contain a question mark. This is because
Lapis can't seem to cope with this, since it thinks the question marks
needs to be interpolated. So instead, we'll drop-down to a lower-level
query mechanism whenever we're manually escaping queries.