Fix api key transition timestamp handling in postgres.
Fix the timestamp parsing on the API backend configuration, for fetching
the transition timestamp details.

Fix the timestamp parsing on the API user fetching, but shifting more of
the logic into the SQL view we're using. This eliminates some of the
parsing logic we were having to do in Lua after fetching the users.