Fixes for redirect rewriting and port handling.
Various fixes around HTTP/HTTPS redirect rewriting behavior from API
backends. Custom ports were previously not being included in the
rewritten redirects. This should now be fixed so the custom ports (as
well as overridden ports are now taken into account).

This was an existing production issue, and not something specific to
this postgres branch we're currently working on.