Rework web app validations to allow for translating field names.
Instead of relying on removing underscores and title-casing the database
column name, rework the validation setup a bit to allow for passing in a
translated, human-readable field name to the validation method. This
also makes validations against he same field slightly less repetitive
(since we now define an array of validations on a single field).