Continue using Elasticsearch v2 as default version, but with v6 compat.
To simplify upgrades for now, we'll continue using Elasticsearch v2 as
the default, bundled version. But the code should all remain compatible
with Elasticsearch v6 (with v2 of the template).

Although, we still need to figure out a way to integrate this into our
test suite so we keep testing v6 and v2 simultaneously.