Redirect all web content to HTTPS by default.
This adjusts the existing config options so that API Umbrella redirects
all web-app and website requests to the HTTPS version by default (while
we had configured most of our instances to do this already, it seems
time to make this the default behavior so there's not the weird
page-specific split in behavior).

It also adds a new configuration option so that any "not found"
responses are redirected to HTTPS.

API behavior remains the same (we default to forcing HTTPS, but not via
redirects).

See https://github.com/18F/api.data.gov/issues/430 for details about the
"not found" redirecting.