Fix cached random tokens potentially not persisting on first start.
The "var/run" directory was not being created prior to the random tokens
being trying to be written. So if this directory didn't already exist
(on initial startup), this could lead to the random tokens changing on
subsequent runs.

To fix this, ensure the directory exists, but also better optimize this
process so new tokens are only generated if needed, and the file is only
written if needed (rather than rewriting it on every startup).