Fix limited admin accounts not being able to publish website backends.
Pending changes for website backends weren't being returned to admin
accounts with limited scopes (only authorized on specific domains). The
reason was that we were inadvertently applying the API backend policy
scope to this permission check instead of the website backend policy (so
the API backend policy would never consider the website backend
authorized).

This fixes that issue and adds more test coverage for the admin
permissions surrounding website backends in the config publishing
process.