Cleanup connection collapsing implementation for TrafficServer v7.
While the behavior might be a bit different from what we had in
TrafficServer v5 and the collapsed_connection plugin, that plugin is
deprecated (and seems to cause errors if used), and the options around
all this have changed a bit. These new tests also cover the connection
collapsing behavior in more detail, testing various situations we
didn't previously test.