Fix elasticsearch index naming, and pull in elasticsearch improvements.
This fixes the elasticsearch index names missing the "v" part of the
"v1" version number in the name after the various version config changes
in this branch.

This also pulls in some of the elasticsearch refactoring to add the
helper library for handling connections and queries (originally on the
lapis-postgres branch).