Improve filesystem permissions and use more restrictive umask.
perp was starting processes with a umask of 0, so various directories
and files had permissions that were too wide open (eg, db directory and
files were world readable/writable). This sets more explicit permissions
everywhere to mitigate this and fix existing installations.

We'll also start using a more restrictive umask everywhere and set the
effective gid of the process to try and have better default permissions.

This also gets rid of some of the recursive "chown -R" commands that
were previously in our startup sequence. This could slow down startup
performance as the number of recursive files grew, and shouldn't be
necessary with the more explicit permissions setup. I think the
recursive command could also possibly lead to race conditions (only seen
in the CI environment, but a situation where the chown failed on a
database file that didn't exist, possibly because the database was being
modified at the same time as the chown).