Fix perp runtool commands if run_args is an empty array.
With the introduction of "set -u" to raise errors on undefined
variables, if run_args is empty (which might be the case if not running
the commands under a different user), then this could cause errors. Use
this slightly funky syntax to not cause errors on empty arrays in bash.