Allow running tests as root, without user-switching to fix lsof tests.
With the changes in 2e595cec2dd748af2638f37e6729b7525f0e3572 to change
the gid of the running process, this broke the ability to run "lsof"
tests in Docker test environments. This is mostly because lsof can't
grab information for processes belonging to different users/groups (even
when running lsof as root) inside a Docker container unless that Docker
container is run in privileged mode (which we can't do in our CircleCI
2.0 environment).

So make the necessary tweaks so that we can always run the tests as the
current user, even if the current user is root.

While in some ways it was nice to test user-switching in our test
environment, in other ways this simply reverts testing back to how we
tested in CircleCI 1.0, where the tests were run as the current "ubuntu"
user (the current user is just root now).

As part of this, also rework the lsof port check tests a bit to get rid
of shell piping so the errors are more obvious on failures, and also
better sanity check the output to ensure we're not missing expected
ports from lsof's output.