Flip the CI cache naming around to prevent conflicts/mismatches.
Since CircleCI's matching is based on prefixes, with the previous
approach it mean the cache for the `updates-trafficserver` branch could
inadvertently get used when building the `updates` branch. So reorder
the naming convention around so conflicts like this aren't as likely
(unless a branch name happened to contain the system arch name).