Remove unused custom DNS and upstream handling (via ngx-dyups).
With the shift to using Traffic Server for DNS resolution and upstream
connections, we no longer need all of this custom code at our nginx
layer. This was some of the thornier things to deal with (DNS
resolution, caching, and configuration of upstreams via the ngx-dyups
plugin), so being able to rely on Traffic Server for this functionality
allows for some nice cleanup.