Fix redirect_https logic.
Skipping redirect rewriting when the redirect_https setting was merely
present wasn't correct, since that skipped rewriting for all responses
when this setting was present, even if they came from the backend.

Instead, tweak things so the redirect rewriting is only skipped when
redirect_https actually triggers a redirect. Also shift how
redirect_https triggers the redirect, to treat it as an error, which
will short-circuite the rest of the middleware, rather than letting them
execute (since in the case of these redirects, we don't need to check
for rate limits, etc).