Replace "run_command" implementation with safer luaposix based one.
Instead of relying on some of the weird workarounds with io:popen to run
shell commands, use a luaposix based approach to executing commands.
This allows us to more cleanly capture stderr, and should generally be
safer, since we don't have to worry about shell escaping a single string
to execute (instead we'll give the command as a table of individual
arguments).