Fix intermittent test failures related to trafficserver restarts.
For the tests that change trafficserver config that require restarts
(DNS tests and keepalive tests), these were randomly failing in the CI
environment. While I wasn't able to precisely reproduce this locally, I
believe the underlying cause was the "sleep 1" after sending the kill
signal to trafficserver. In some cases, this wasn't long enough, which I
believe explains the CI failures, since essentially the tests were
taking place without trafficserver having actually restarted (so the
expected configuration was not actually in effect).

This attempts to fix it by more properly waiting for the trafficserver
process to restart, by ensuring the PID actually changes before
proceeding.

This also better abstracts some of the similar perpctl commands in the
test suite, so other restarts and signals are better handled.