Fix console output so all services and access logs get output to stdout
Fixes so nginx access logs also get output to stdout. Trafficserver's
logging is a bit trickier (since the symlink trick doesn't work there),
so instead we'll try redirecting the log file output with rsyslog.

In testing, the imfile handling in rsyslog doesn't perfectly handle
rotation (some go missing on rotation before the new file is picked up
again), so this isn't a perfect solution, but since the trafficserver
access logs mostly aid in debugging, we'll go with this for now instead
of introducing more new dependencies.