Fix nginx duplicate server name warnings in certain cases.
If the api-umbrella.yml config manually defined a wildcard host, then
two hosts could end up defined for the wildcard server name ("_").

This cleans up the implementation a bit, and prevents this duplicate
server name warning. This doesn't actually change the existing behavior,
so there's still some certain configuration situations that might not
really make sense (non-default wildcards), but we'll keep everything
defined as-is to match the hosts config.