Fix web-app so it receives proper Host header. Cleanup host handling.
With the switch to treating the web-app as just another API backend, the
"Host" header it was seeing was 127.0.0.1, which led to incorrect URLs
being generated by the backend (eg, for Google Auth redirect_uri). Fix
this by not setting an explicit backend host header, so the original
host header flows through to the web-app.

This led to an overall cleanup and standardization of detecting the
forwarded proto and port information, as well as redirect handling. Now
that all requests (APIs, website backends, and the web-app) are flowing
through the same lua stack for proxying, we no longer need some of the
special nginx config that we previously had in place for detecting some
of these variables and cleaning up redirects.

These changes lead to a few small changes in the redirect rewriting
behavior for the website backend and web-app backend responses, but
these changes shouldn't impact API backend rewriting. And the changes
actually lead to more consistent handling across all the proxying.