Add ability to perform HTTPS redirects on API URLs based on regex.
The use case for this is to primarily redirect the root "/" path to
HTTPS when an API backend is setup on the root path. For most API
backend handling, we want to return our 400 error for HTTP requests
(since API clients may not follow redirects), but this may be useful on
the root path for compliance purposes.