Fix repeated edits in development mode.
unloadAll seems to be behaving differently after the Ember upgrade,
which was leading to problems in development mode editing the same
record repeatedly (but strangely not in production builds). Fix this by
wrapping the unloadAll in the run loop. Add better test coverage of the
various things this UncachedModel mixin handles.