Fix race conditions in tests with web app reloading.
Some config changes weren't being picked up in tests due to the old
nginx web app processes still being hit (before the new config was in
place). So improve the reload process for both the router and web app
nginx processes so that during tests, we always wait for the new nginx
workers to be active.