Always reload web app to keep config in sync.
With the changes in 8486fa437b6ce10624acb0b014fa26221b0fbdd1, we need to
ensure the web app is reloaded and picks up file-based config changes on
reloads too. So simplify the reloading process by always reloading the
web app when the router is reloaded. Since the web app should now reload
pretty quickly, it shouldn't be a big deal to also reload it and keep
everything better in sync.