Fix some race conditions leading to mismatched config in tests env.
When the "api-umbrella health" command was run after "api-umbrella run"
(but before run had really started), then it could lead to both
processes trying to generate a new secret key and caching it. This, in
turn, could lead to random test failures, since the secret key used for
encryption was being overwritten but the second process.

This fixes this behavior by only writing the generated values if the
overall config writing mode is enabled, which is usually just the
responsibility of the "api-umbrella run" command.