Fix intermittent test failures with parallel tests.
The usage of the "force_publish_config" method previously wasn't safe to
call in parallel tests, since it could lead to the configuration
overrides replace the config with an older version of the config. This
fixes it, so the method can still be used in parallel tests by better
wrapping the method's usage in a lock to ensure it doesn't conflict with
other config publishing.