Switch to resty-moongoo for Mongo to Postgres migration script.
The lua-mongo library stopped working within the context of the "resty"
CLI. This was working, so the only thing I can think of would be the
switch to Ubuntu 18.04. Requiring the lua-mongo library would simply
abort the resty CLI process. This problem did not occur if calling
things with luajit, so it somehow seemed to be impacted by running
inside nginx. lua-mongo uses blocking APIs (although we weren't really
worried about this for just a one-off migration script), so something
weird seems to be happening.

In order to workaround this, switch to resty-moongoo, which seems to
work without issue in OpenResty, and I think at least supports all the
functionality we need in order to perform our migration.