Add automatic json wrapping behavior to JSON APIs.
When accepting a JSON body for parameters, also accept the attributes as
root attributes, rather than inside the normal nested area. While this
inconsistency in API doesn't seem ideal (the form encoded params don't
behave the same way), in order to maintain compatibility with the Rails
API (which has this behavior by default for JSON APIs), we'll replicate
it here.