Fix DNS stale test behavior for Trafficserver 8.
In Trafficserver 8, the behavior for stale DNS usage has changed
slightly. If a valid NXDOMAIN result is returned, then the stale record
is expired and will no longer be used. Instead, stale records are only
used if the lookup actually fails. This new behavior seems fine (just
different), So we need to change our test suite to account for these
differences.