Improve falling back to Origin header for enforcing required referrers.
If an API or API key was setup with referer requirements, but a browser
did not send a Referer header, then we were falling back to using the
Origin header for matching the referer requirements. However, the
previous approach did not take into account that the Origin header does
not include the path component of the URL, so depending on how the
referer requirements were constructed, the Origin-based approach may
not have matched.

This loosens the matching requirements when just an Origin header is
present to better deal with situations where Referers are stripped, but
the Origin remains. These seem to generally represent real browser
requests, and not forgery attempts. Since the referer requirements are
mainly meant to provide basic protection against key reuse in browsers
(and aren't foolproof for security purposes), loosening these
requirements seems acceptable and should ensure legitimate users hitting
the edge-cases aren't affected.