Statically link pcre and openssl to openresty.
This is how we were previously building things, but due to the PCRE bug
in OpenResty 1.15.8.1, we ended up switching things to not be statically
linked. However, this has caused more issues with the version of OpenSSL
installed for OpenResty fighting with the system OpenSSL. So, for
example, this was breaking the installation of "lua-resty-auto-ssl" via
luarocks, since the "curl" command that was part of the luarock install
process was trying to use our newer version of OpenSSL, but curl was
linked to the system version of OpenSSL.

So by going back to statically linked for these dependencies, this
should hopefully resolve these conflicts in an easier way.