Fixes for new trafficserver timeout behavior.
Trafficserver 9's connect timeout is now more of a proper connect
timeout, so adjust various things to account for this.

GET requests are still retried, so patch for that behavior (maybe
revisit this to disable retries altogether or allow them as is the
default behavior, if this ends up over complicating things).

Fix an issue with the cache key not always being generated correctly,
since I think it needs to be a URL.