Allow limited admins to view API keys they create themselves.
This was lost in dc888eccaa9bc5acdb27a2d7380c6d9d5bd9e6fd. This brings
this back, but without the more complicated time or role aspect. We will
assume that if the admin was able to create the user, then it would only
have roles they had permissions to. If a superuser then adds other roles
to the user, then the limited admin will continue to be able to view the
api key, but only for the keys they had originally created themselves.