Move all templates to to etlua templates.
I had started this migration with some of the files, but to simplify
things and so we're not supporting two different template formats (and
with some of mustache's oddities), finish this migration so everything's
standardized.

Most of these changes were automated, so there may be some issues to
discover, though.