Begin work on more efficient, less resource-heavy rate limiting.
Simplify rate limiting strategy to use an estimated counting approach,
which significantly reduces number of separate counter buckets windows
used and removes the need for our "accuracy" settings and windows.

Cache when a rate limit has been exceeded, so no further querying or
counting is needed for the duration of the rate limit being exceeded.

Not fully tested or functional yet, but initial benchmarks look like
this is both quite a bit faster and more memory efficient than the
previous implementation. The code is also hopefully simpler.