Tag and release in separate workflow.
So that the tagging and release process doesn't necessarily need to wait
for tests if they've already succeeded.

Also try to improve caching of tagged builds so they can leverage other
recent builds rather than having a per-tag cache.

Rearrange tagging naming scheme so it's easier to identify the tags
related to a single branch.