Solved periodic CI failure.
The TRUNCATE statement was cascading to the rate_limits table and delete
all contents (not just rate limits associated with API backends), so
that's why the user rate limits were disappearing if this test was ran
first. So fix this so we no longer do a truncate, but instead just reset
the sequence that's actually intended to be tested.