Fix version bumps to Crane not downloading new file.
Since the crane downloads didn't contain the version number in the
filename (only further up in the URL path), then when bumping Crane, new
downloads weren't being performed. This fixes it by including a hash of
the URL in the download filename (and then symlinking it to the original
path for compatibility).