Consistent lua nested dependencies. Downgrade pgmoon.
A long standing issue with our Lua dependencies is that nested
dependency versions weren't pinned. This meant that a fresh installation
may install newer versions of nested dependencies than what we last
tested with. This recently bit us with pgmoon being upgraded (as a
dependency of lapis) from 1.14 to 1.15, which broke some things.

This reworks all of our Lua dependencies to take advantage of the lock
files that LuaRocks now supports to better deal with this. This should
ensure all Lua dependencies and their nested dependencies are pinned in
a lock file.

In the lock file rollback pgmoon to 1.14 for current compatibility, but
will revisit dealing with newer version compatibility.

Other semi-related changes this led to:

- Update build/install process to use the LuaRocks lock file to install
  all of the LuaRocks at once (more akin to how we install Yarn or Gem
  dependencies, rather than treating every Lua dependency as a separate
  package).
- Remove usage of OPM in favor of LuaRocks everywhere.
- Update the "outdated" task to handle looking for outdated LuaRocks
  dependencies using the new lock file.
- Refactor the "outdated" task as a Thor class, instead of a mixed Rake
  and plain Ruby class.
- Shift over Rake tasks to Thor for easier management of tasks written
  in Ruby.
- Remove icu4c dependency since this was only used for RHEL7 (otherwise
  we use the system packages), which I don't think we need to support
  anymore.
- Remove mongo dependencies. We were installing some Lua mongo libraries
  for the Mongo to Postgres data migration script, but remove these for
  now, and we'll revisit the data migration script later.