Alternative implementation of sorted JSON output.
See if this fixes the mysterious CI failures we've been seeing. They
seem to have cropped up around the same time we introduced the rapidjson
dependency, with some of our json output being malformed (arrays being
output as hashes with -1 as the starting hash key), and the
nginx-web-app failing to startup altogether with core dumps. Not exactly
sure why the rapidjson dependency would lead to this, but we'll see...

This reverts commit f1098e75227b0f9f285d878cdbfb889a5afa17aa.