Fix custom error messages not working when sub-url settings present
This reworks how the settings all get merged and handled to hopefully
make all of this more logical and performant. We now merge all of the
default settings, api settings, and sub-settings together when the
config is loaded (rather than at runtime per request). This should make
aspects more performant, since we're performing less work when
processing requests.

This also attempts to simplify various funky things in our merging logic
that had cropped up over the years so the merging logic is more
straightforward and consistent (so we're not treating sub-url settings
differently).

Add test coverage for some of the sub-url stuff that was misbehaving
(and add coverage for some sub-url stuff in other areas like rate limits
that was working, but we didn't have test coverage for).