Update Trafficserver and update some tests to accommodate.
Trafficserver 9.2+ doesn't seem to work with our previous testing
approach of returning HTTP response headers while the request body is
still being streamed in. This was an admittedly weird setup to try and
test the request and response streaming timeouts separately. Since
request bodies still streaming after the response headers have been sent
doesn't seem like a particularly realistic scenario we need to support,
this changes how we test request vs response timeouts to try and
better separate them so we don't need this funky setup. Now we set
different timeouts as needed to test the different scenarios, rather
than relying on this weird workaround for static timeouts.

Also fix some issues with the X-Cache header after the upgrade (explored
doing this directly in Trafficserver, but that still has issues, but
committing for reference an approach that I think could work in the
future).