Prevent duplicative API key signup submissions.
The form being disabled wasn't happening until after recaptcha stuff
occurred, so if a user double/triple clicked on the submit button, it
could lead to duplicate signups. Fix this by disabling the submit button
immediately. This requires handling the possibility of the user closing
the recaptcha prompt, which requires a bit of a workaround.

Also add some clarifying comments about some areas to revisit once we
enforce captchas globally.